server/requirements.txt
fastapi
uvicorn

python-dotenv

python-multipart

pydantic

pinecone
pinecone-client

sentence-transformers

groq

numpy
